Trekkjer ut dei forskjellige Unleash-implementasjonane i eigne klassar, synes det er enklare å få oversikt da
