+++ src/main/kotlin/no/nav/familie/ba/sak/config/FeatureToggleConfig.kt
@@ -1,11 +1,8 @@
 package no.nav.familie.ba.sak.config
 
-import io.getunleash.DefaultUnleash
-import io.getunleash.UnleashContext
-import io.getunleash.UnleashContextProvider
-import io.getunleash.strategy.GradualRolloutRandomStrategy
 import io.getunleash.strategy.Strategy
-import io.getunleash.util.UnleashConfig
+import no.nav.familie.ba.sak.config.featureToggle.DummyFeatureToggleService
+import no.nav.familie.ba.sak.config.featureToggle.UnleashFeatureToggleService
 import no.nav.familie.ba.sak.sikkerhet.SikkerhetContext
 import org.slf4j.LoggerFactory
 import org.springframework.boot.context.properties.ConfigurationProperties
@@ -39,32 +36,7 @@ class FeatureToggleConfig(
             lagDummyFeatureToggleService()
         }
 
-    private fun lagUnleashFeatureToggleService(): FeatureToggleService {
-        val defaultUnleash = DefaultUnleash(
-            UnleashConfig.builder()
-                .appName(unleash.applicationName)
-                .unleashAPI(unleash.uri)
-                .unleashContextProvider(lagUnleashContextProvider())
-                .build(),
-            ByClusterStrategy(unleash.cluster),
-            ByAnsvarligSaksbehandler(),
-            GradualRolloutRandomStrategy()
-        )
-
-        return object : FeatureToggleService {
-            override fun isEnabled(toggleId: String, defaultValue: Boolean): Boolean {
-                return defaultUnleash.isEnabled(toggleId, defaultValue)
-            }
-        }
-    }
-
-    private fun lagUnleashContextProvider(): UnleashContextProvider {
-        return UnleashContextProvider {
-            UnleashContext.builder()
-                .appName(unleash.applicationName)
-                .build()
-        }
-    }
+    private fun lagUnleashFeatureToggleService(): FeatureToggleService = UnleashFeatureToggleService(unleash)
 
     class ByClusterStrategy(private val clusterName: String) : Strategy {
 
@@ -87,17 +59,7 @@ class FeatureToggleConfig(
         override fun getName(): String = "byAnsvarligSaksbehandler"
     }
 
-    private fun lagDummyFeatureToggleService(): FeatureToggleService {
-        return object : FeatureToggleService {
-            override fun isEnabled(toggleId: String, defaultValue: Boolean): Boolean {
-                if (unleash.cluster == "lokalutvikling") {
-                    return false
-                }
-
-                return defaultValue
-            }
-        }
-    }
+    private fun lagDummyFeatureToggleService(): FeatureToggleService = DummyFeatureToggleService(unleash)
 
     companion object {
         const val KAN_DIFFERANSEBEREGNE_SØKERS_YTELSER = "familie-ba-sak.differanseberegn-sokers-ytelser"
